↩[Move] model of posts to folder of all models
